requirement_new.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
attrs==25.4.0
beautifulsoup4==4.14.2
brotli==1.2.0
bs4==0.0.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
comm==0.2.3
dataclasses-json==0.6.7
dateparser==1.2.2
ddgs==9.9.0
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
duckduckgo_search==8.1.1
executing==2.2.1
fastapi==0.121.1
fastuuid==0.14.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.10.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
langchain==1.0.5
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.4
langchain-huggingface==1.0.1
langchain-litellm==0.3.0
langchain-openai==1.0.2
langchain-text-splitters==1.0.0
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.42
lite-llm-client==0.2.0
litellm==1.79.3
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib-inline==0.2.1
multidict==6.7.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.3.4
openai==2.7.2
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
platformdirs==4.5.0
primp==0.15.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.3
pure_eval==0.2.3
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pymilvus==2.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.28.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8
SQLAlchemy==2.0.44
stack-data==0.6.3
starlette==0.49.3
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
